+++ practicum_work2/4/LomutoQsort.cpp
@@ -6,16 +6,16 @@
 #include "LomutoQsort.h"
 #include "../Common.h"
 
-static void Qsort(int* arr, int left, int right);
+static void Qsort(int* arr, size_t left, size_t right);
 
-static int Partition(int* arr, int left, int right)
+static size_t Partition(int* arr, int left, int right)
 {
     assert(arr != NULL);
 
     int pivot = arr[left + (right - left) / 2];
     swap(&arr[right], &arr[left + (right - left) / 2], sizeof(int));
-    int i = left;
-    for (int j = left; j <= right; ++j)
+    size_t i = left;
+    for (size_t j = left; j <= right; ++j)
     {
         if (arr[j] < pivot)
         {
@@ -26,13 +26,13 @@ static int Partition(int* arr, int left, int right)
     return i;
 }
 
-static void Qsort(int* arr, int left, int right)
+static void Qsort(int* arr, size_t left, size_t right)
 {
     assert(arr != NULL);
 
     if (left < right)
     {
-        int piv_idx = Partition(arr, left, right);
+        size_t piv_idx = Partition(arr, left, right);
         Qsort(arr, left, piv_idx - 1);
         Qsort(arr, piv_idx + 1, right);
     }
@@ -43,5 +43,5 @@ void LomutoQsort(int* arr, size_t size)
     assert(arr != NULL);
     assert(size > 0);
 
-    Qsort(arr, 0, (int)size - 1);
+    Qsort(arr, (size_t)0, size - 1);
 }
