+++ practicum_work2/4/OneBranchHoaraQsort.cpp
@@ -7,16 +7,16 @@
 
 #include "OneBranchHoaraQsort.h"
 
-static void Qsort(int* arr, int left, int right);
+static void Qsort(int* arr, size_t left, size_t right);
 
-static int Partition(int* arr, int left, int right)
+static size_t Partition(int* arr, size_t left, size_t right)
 {
     assert(arr != NULL);
 
-    int piv_idx = left + (right - left) / 2;
+    size_t piv_idx = left + (right - left) / 2;
     int pivot   = arr[piv_idx];
-    int i       = left;
-    int j       = right;
+    size_t i       = left;
+    size_t j       = right;
     while(i <= j)
     {
         while(arr[i] < pivot)
@@ -36,13 +36,13 @@ static int Partition(int* arr, int left, int right)
     return j;
 }
 
-static void Qsort(int* arr, int left, int right)
+static void Qsort(int* arr, size_t left, size_t right)
 {
     assert(arr != NULL);
 
     while (left < right)
     {
-        int piv_idx = Partition(arr, left, right);
+        size_t piv_idx = Partition(arr, left, right);
 
         if (piv_idx - left < right - piv_idx)
         {
@@ -62,5 +62,5 @@ void OneBranchHoaraQsort(int* arr, size_t size)
     assert(size > 0);
     assert(arr != NULL);
 
-    Qsort(arr, 0, (int)size - 1);
+    Qsort(arr, (size_t)0, size - 1);
 }
