+++ practicum_work2/5/Median3Qsort.cpp
@@ -6,7 +6,7 @@
 #include "MedianQsort.h"
 #include "../Common.h"
 
-static void Qsort(int* arr, int left, int right);
+static void Qsort(int* arr, size_t left, size_t right);
 
 int GiveMiddle(int first, int second, int third) 
 {
@@ -24,13 +24,12 @@ int GiveMiddle(int first, int second, int third)
     }
 }
 
-
-static int Partition(int* arr, int left, int right)
+static size_t Partition(int* arr, size_t left, size_t right)
 {
-    int piv_idx = left + (right - left) / 2;
-    int pivot   = GiveMiddle(arr[left], arr[piv_idx], arr[right]);
-    int i       = left;
-    int j       = right;
+    size_t piv_idx = left + (right - left) / 2;
+    int pivot      = GiveMiddle(arr[left], arr[piv_idx], arr[right]);
+    size_t i       = left;
+    size_t j       = right;
     while(i <= j)
     {
         while(arr[i] < pivot)
@@ -50,11 +49,11 @@ static int Partition(int* arr, int left, int right)
     return j;
 }
 
-static void Qsort(int* arr, int left, int right)
+static void Qsort(int* arr, size_t left, size_t right)
 {
     if (left < right)
     {
-        int piv_idx = Partition(arr, left, right);
+        size_t piv_idx = Partition(arr, left, right);
         Qsort(arr, left, piv_idx);
         Qsort(arr, piv_idx + 1, right);
     }
