+++ practicum_work2/5/Median3RandQsort.cpp
@@ -6,7 +6,7 @@
 #include "Median3RandQsort.h"
 #include "../Common.h"
 
-static void Qsort(int* arr, int left, int right);
+static void Qsort(int* arr, size_t left, size_t right);
 
 int GetMedian(int first, int second, int third) 
 {
@@ -25,15 +25,15 @@ int GetMedian(int first, int second, int third)
 }
 
 
-static int Partition(int* arr, int left, int right)
+static size_t Partition(int* arr, size_t left, size_t right)
 {
     assert(arr != NULL);
 
     int pivot   = GetMedian(arr[left + (rand() % (right - left + 1))], 
                             arr[left + (rand() % (right - left + 1))], 
                             arr[left + (rand() % (right - left + 1))]);
-    int i       = left;
-    int j       = right;
+    size_t i       = left;
+    size_t j       = right;
     while(i <= j)
     {
         while(arr[i] < pivot)
@@ -53,13 +53,13 @@ static int Partition(int* arr, int left, int right)
     return j;
 }
 
-void Qsort(int* arr, int left, int right)
+void Qsort(int* arr, size_t left, size_t right)
 {
     assert(arr != NULL);
 
     if (left < right)
     {
-        int piv_idx = Partition(arr, left, right);
+        size_t piv_idx = Partition(arr, left, right);
         Qsort(arr, left, piv_idx);
         Qsort(arr, piv_idx + 1, right);
     }
@@ -70,5 +70,5 @@ void Median3RandQsort(int* arr, size_t size)
     assert(size > 0);
     assert(arr != NULL);
 
-    Qsort(arr, 0, (int)size - 1);
+    Qsort(arr, (size_t)0, size - 1);
 }
