+++ practicum_work2/5/MedianQsort.cpp
@@ -6,16 +6,16 @@
 #include "MedianQsort.h"
 #include "../Common.h"
 
-static void Qsort(int* arr, int left, int right);
+static void Qsort(int* arr, size_t left, size_t right);
 
-static int Partition(int* arr, int left, int right)
+static size_t Partition(int* arr, size_t left, size_t right)
 {
     assert(arr != NULL);
 
-    int piv_idx = left + (right - left) / 2;
-    int pivot   = arr[piv_idx];
-    int i       = left;
-    int j       = right;
+    size_t piv_idx = left + (right - left) / 2;
+    int pivot      = arr[piv_idx];
+    size_t i       = left;
+    size_t j       = right;
     while(i <= j)
     {
         while(arr[i] < pivot)
@@ -35,13 +35,13 @@ static int Partition(int* arr, int left, int right)
     return j;
 }
 
-static void Qsort(int* arr, int left, int right)
+static void Qsort(int* arr, size_t left, size_t right)
 {
     assert(arr != NULL);
 
     if (left < right)
     {
-        int piv_idx = Partition(arr, left, right);
+        size_t piv_idx = Partition(arr, left, right);
         Qsort(arr, left, piv_idx);
         Qsort(arr, piv_idx + 1, right);
     }
@@ -52,5 +52,5 @@ void MedianQsort(int* arr, size_t size)
     assert(size > 0);
     assert(arr != NULL);
 
-    Qsort(arr, 0, (int)size - 1);
+    Qsort(arr, (size_t)0, size - 1);
 }
